tests/requirements.txt
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
PyJWT==2.8.0
passlib==1.7.4
boto3==1.34.0
moto==4.2.11
hypothesis==6.92.0
